22/07 update SQLite Put and Delete 2

+++ app/src/main/java/com/tuncaksoy/catschallenge/viewmodel/HomePageViewModel.kt
@@ -2,6 +2,7 @@ package com.tuncaksoy.catschallenge.viewmodel
 
 import android.app.Application
 import android.content.pm.ChangedPackages
+import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.tuncaksoy.catschallenge.model.Cats
@@ -17,28 +18,32 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import java.util.ArrayList
 
 open class HomePageViewModel(application: Application) : BaseViewModel(application) {
     val cats = MutableLiveData<List<Cats>>()
     private val CatApiServis = CatAPIServis()
     private val disposable = CompositeDisposable()
     private val catSharedPreferences = CatSharedPreferences(getApplication())
-    var positionId : Int? = null
-    var changedPositon :Int? = null
-    var positionGenus : String? = null
+    var positionId: Int? = null
+    var changedPositon: Int? = null
+    var positionGenus: String? = null
+    var positionFavorites: Boolean? = null
 
-    private val _catss = MutableStateFlow(HomeUiState(onFavoritesChanged = { catId, catGenus->
-        viewModelScope.launch {
-            positionGenus = catGenus
-            positionId = catId
-            println(positionId)
-            println(changedPositon)
+    private val _catss =
+        MutableStateFlow(HomeUiState(onFavoritesChanged = { catId, catGenus, catFavorites ->
+            viewModelScope.launch {
+                positionGenus = catGenus
+                positionId = catId
+                positionFavorites = catFavorites
+                println(positionId)
+                println(changedPositon)
                 getDataAPI()
 
-        }
-    }))
+            }
+        }))
 
-    val catss : StateFlow<HomeUiState> = _catss.asStateFlow()
+    val catss: StateFlow<HomeUiState> = _catss.asStateFlow()
 
 
     fun refreshData() {
@@ -53,9 +58,9 @@ open class HomePageViewModel(application: Application) : BaseViewModel(applicati
 
                     override fun onSuccess(t: List<Cats>) {
                         if (positionId != null && positionGenus != null) {
-                                if (changedPositon != positionId) {
-                                    putSQLITE(t, positionId!!, positionGenus!!)
-                                    changedPositon = positionId
+                            if (changedPositon != positionId) {
+                                putSQLITE(t, positionId!!, positionGenus!!, positionFavorites!!)
+                                changedPositon = positionId
                             }
                         }
                         showCats(t)
@@ -73,13 +78,33 @@ open class HomePageViewModel(application: Application) : BaseViewModel(applicati
     }
 
 
-    fun putSQLITE(catsList: List<Cats>,catId : Int,catGenus : String) {
+    fun putSQLITE(catsList: List<Cats>, catId: Int, catGenus: String, catFavorites: Boolean) {
         viewModelScope.launch(Dispatchers.IO) {
             val dao = CatDatabase(getApplication()).catDao()
+            Log.d("selam","catGenus" + catGenus)
+            val favoritesList = dao.getAllGenus()
+            var catFavoritess = false
+            if (favoritesList.size!=0){
+                var i = 0
+                while(i<favoritesList.size) {
+                    var favoritesGenus = favoritesList[i].catGenus
+                    i++
+                    Log.d("selam","favoritesGenusif->"+favoritesGenus)
+                    if (catGenus == favoritesGenus) {
+                        catFavoritess = true
+                    }
+                }
+            }
+            Log.d("selam","CatFavorites->" + catFavoritess.toString())
+            //println(favoritesList)
             //dao.deleteAllCat()
-            dao.insert(catsList[catId])
-            dao.deleteCat(catGenus)
-            println(catsList[catId])
+            if (catFavoritess == true) {
+                dao.deleteCat(catGenus)
+                println("yokburada")
+            } else {
+                dao.insert(catsList[catId])
+                println("burada")
+            }
             //dao.insertAll(cats)
             /*val uuidListesi = dao.insertAll(*catsList.toTypedArray())
             var i = 0
@@ -94,5 +119,5 @@ open class HomePageViewModel(application: Application) : BaseViewModel(applicati
 }
 
 data class HomeUiState(
-    val onFavoritesChanged: (Int?, String?) -> Unit
+    val onFavoritesChanged: (Int?, String?, Boolean?) -> Unit
 )
